fix: switch to bf16 for HF Jobs (newer transformers + bnb prefers bf16, no GradScaler needed)

+++ training/train_attacker.py
@@ -160,7 +160,7 @@ def train_attacker(
         model_name,
         quantization_config=bnb_config,
         device_map="auto",
-        dtype=torch.float16,
+        dtype=torch.bfloat16,
     )
     model = prepare_model_for_kbit_training(model)
 
@@ -198,7 +198,7 @@ def train_attacker(
         num_generations=tr["rollouts_per_episode"],
         temperature=tr["temperature"],
         top_p=tr["top_p"],
-        fp16=True,
+        bf16=True,
     )
 
     if opponent is not None:
